emulator: fix issue where emulator would exit prematurely
Using a 32 bit number to count instructions executed hits the limit
surprisingly fast. Expand max instruction count to 64 bits.

+++ tools/emulator/core.c
@@ -322,9 +322,9 @@ bool stoppedOnFault(const Core *core)
     return core->crashed;
 }
 
-bool executeInstructions(Core *core, uint32_t threadId, uint32_t totalInstructions)
+bool executeInstructions(Core *core, uint32_t threadId, uint64_t totalInstructions)
 {
-    uint32_t instructionCount;
+    uint64_t instructionCount;
     uint32_t thread;
 
     core->singleStepping = false;
